feat(EMS-1181): declarations - move saveData and saveAndBack functions/controllers into generic declarations route
